here's my data scraper for north carolina
saved text files to hard drive 1-by-1 and searched through each letter of the alphabet to complete the task.  Data format was difficult to navigate in that it seemed to be formatted for mailing labels.  A better version of this will have contingencies built in for testing, qualifying, and rendering data in fewer passes.  I'm learning, so forgive the simplicity and likely silliness of some of this code.